+++ lib/asset_ram.rb
@@ -5,9 +5,9 @@ require_relative "asset_ram/version"
 #
 # Use in views to cache the asset path computation.
 #
-# For example, in a HAML file:
+# Preferred usage (since v1.1.0):
 #
-# = AssetRam::Helper.cache { favicon_link_tag('favicon/favicon.ico', rel: 'icon') }
+#   = AssetRam.cache { favicon_link_tag('favicon/favicon.ico', rel: 'icon') }
 #
 # The calculated asset paths are keyed by source file name and line number.
 # The results are stored in RAM.
@@ -15,7 +15,11 @@ require_relative "asset_ram/version"
 # Sometimes, a key is needed if the code is run in different contexts, like
 # a multi-tenant site:
 #
-# = AssetRam::Helper.cache(key: site) { stylesheet_link_tag("themes/#{site}", media: nil) }
+#   = AssetRam.cache(key: site) { stylesheet_link_tag("themes/#{site}", media: nil) }
+#
+# For compatibility, you can still use:
+#
+#   = AssetRam::Helper.cache { ... }
 #
 # To test and compare if this lib actually improves performance,
 # set the ASSET_RAM_DISABLE env var and it will transparently never cache.
@@ -25,14 +29,22 @@ module AssetRam
   class Error < StandardError; end
 
   ##
+  # The simpler API: AssetRam.cache { ... }
+  #
+  def self.cache(key: '', &blk)
+    Helper.cache(key: key, &blk)
+  end
+
+  ##
   # Our own asset helper which memoizes Rails' asset helper calls.
+  #
   class Helper
     @@_cache = {}
 
 
     def self.cache(key: '', &blk)
       cache_key = blk.source_location
-      cache_key << key if key.present?
+      cache_key << key if !key.to_s.empty?
 
       cache_by_key(cache_key, &blk)
     end